requirements.txt
numpy>=1.13.1
git+ssh://git@github.com/readbeyond/aeneas.git@4525d84f30d374f4078b84234e5ee49db104b52f
pytest
click==6.7
typing==3.6.1
num2words==0.5.7
roman==2.0.0
nltk==3.3
python-vlc==3.0.102